+++ src/cli/configuration_builder.js
@@ -2,9 +2,11 @@ import _ from 'lodash'
 import ArgvParser from './argv_parser'
 import fs from 'mz/fs'
 import path from 'path'
-import PathExpander from './path_expander'
 import OptionSplitter from './option_splitter'
-import Promise from 'bluebird'
+import Promise, { promisify } from 'bluebird'
+import glob from 'glob'
+
+const globP = promisify(glob)
 
 export default class ConfigurationBuilder {
   static async build(options) {
@@ -14,7 +16,6 @@ export default class ConfigurationBuilder {
 
   constructor({ argv, cwd }) {
     this.cwd = cwd
-    this.pathExpander = new PathExpander(cwd)
 
     const parsedArgv = ArgvParser.parse(argv)
     this.args = parsedArgv.args
@@ -29,13 +30,13 @@ export default class ConfigurationBuilder {
     let supportCodePaths = []
     if (!listI18nKeywordsFor && !listI18nLanguages) {
       featurePaths = await this.expandFeaturePaths(unexpandedFeaturePaths)
-      const featureDirectoryPaths = this.getFeatureDirectoryPaths(featurePaths)
-      const unexpandedSupportCodePaths =
-        this.options.require.length > 0
-          ? this.options.require
-          : featureDirectoryPaths
-      supportCodePaths = await this.expandSupportCodePaths(
-        unexpandedSupportCodePaths
+      let unexpandedSupportCodePaths = this.options.require
+      if (unexpandedSupportCodePaths.length === 0) {
+        unexpandedSupportCodePaths = this.getFeatureDirectoryPaths(featurePaths)
+      }
+      supportCodePaths = await this.expandPaths(
+        unexpandedSupportCodePaths,
+        '.js'
       )
     }
     return {
@@ -45,6 +46,7 @@ export default class ConfigurationBuilder {
       formatOptions: this.getFormatOptions(),
       listI18nKeywordsFor,
       listI18nLanguages,
+      parallel: this.options.parallel,
       profiles: this.options.profile,
       pickleFilterOptions: {
         featurePaths: unexpandedFeaturePaths,
@@ -59,15 +61,33 @@ export default class ConfigurationBuilder {
         worldParameters: this.options.worldParameters
       },
       shouldExitImmediately: !!this.options.exit,
-      supportCodePaths
+      supportCodePaths,
+      supportCodeRequiredModules: this.options.requireModule
     }
+  }
+
+  async expandPaths(unexpandedPaths, defaultExtension) {
+    const expandedPaths = await Promise.map(
+      unexpandedPaths,
+      async unexpandedPath => {
+        const matches = await globP(unexpandedPath, {
+          absolute: true,
+          cwd: this.cwd
+        })
+        return Promise.map(matches, async match => {
+          if (path.extname(match) === '') {
+            return globP(`${match}/**/*${defaultExtension}`)
+          }
+          return match
+        })
+      }
+    )
+    return _.flattenDepth(expandedPaths, 2).map(x => path.normalize(x))
   }
 
   async expandFeaturePaths(featurePaths) {
     featurePaths = featurePaths.map(p => p.replace(/(:\d+)*$/g, '')) // Strip line numbers
-    return this.pathExpander.expandPathsWithExtensions(featurePaths, [
-      'feature'
-    ])
+    return this.expandPaths(featurePaths, '.feature')
   }
 
   getFeatureDirectoryPaths(featurePaths) {
@@ -124,19 +144,6 @@ export default class ConfigurationBuilder {
         return featurePaths
       }
     }
-    return ['features']
-  }
-
-  async expandSupportCodePaths(supportCodePaths) {
-    const extensions = ['js']
-    this.options.compiler.forEach(compiler => {
-      const [extension, module] = OptionSplitter.split(compiler)
-      extensions.push(extension)
-      require(module)
-    })
-    return this.pathExpander.expandPathsWithExtensions(
-      supportCodePaths,
-      extensions
-    )
+    return ['features/**/*.feature']
   }
 }