+++ src/cli/configuration_builder_spec.js
@@ -33,8 +33,9 @@ describe('Configuration', () => {
         formats: [{ outputTo: '', type: 'progress' }],
         listI18nKeywordsFor: '',
         listI18nLanguages: false,
+        parallel: 0,
         pickleFilterOptions: {
-          featurePaths: ['features'],
+          featurePaths: ['features/**/*.feature'],
           names: [],
           tagExpression: ''
         },
@@ -47,7 +48,8 @@ describe('Configuration', () => {
           worldParameters: {}
         },
         shouldExitImmediately: false,
-        supportCodePaths: []
+        supportCodePaths: [],
+        supportCodeRequiredModules: []
       })
     })
   })